app/serializers.py
from rest_framework import serializers
from .models import Product, Customer, Order


class ProductSerializer(serializers.ModelSerializer):
    class Meta:
        model = Product
        fields = '__all__'


class CustomerSerializer(serializers.ModelSerializer):
    class Meta:
        model = Customer
        fields = '__all__'


class OrderSerializer(serializers.ModelSerializer):
    total_amount = serializers.SerializerMethodField()

    class Meta:
        model = Order
        fields = ['id', 'customer', 'product', 'quantity', 'total_amount']

    def get_total_amount(self, obj):
        return obj.total_amount()
